Refactor file structure and rename files
- Updated directory structure for better organization.
- Renamed files to improve clarity and consistency.
