+++ vampirenumber/Main.java
@@ -1,8 +1,8 @@
-package Assignment6;
+package assignment6.vampirenumber;
 
 import java.util.Arrays;
 
-public class VampireNumbers {
+public class Main {
 
     public static void main(String[] args) {
         int count = 0;
